refactor: move reactiontype to profile package
