+++ src/main/java/com/meoguri/linkocean/domain/bookmark/service/dto/GetDetailedBookmarkResult.java
@@ -6,7 +6,7 @@ import java.util.Map;
 
 import com.meoguri.linkocean.domain.bookmark.entity.vo.Category;
 import com.meoguri.linkocean.domain.bookmark.entity.vo.OpenType;
-import com.meoguri.linkocean.domain.bookmark.entity.vo.ReactionType;
+import com.meoguri.linkocean.domain.profile.command.entity.vo.ReactionType;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
