+++ src/test/java/com/meoguri/linkocean/domain/profile/command/persistence/ProfileRepositoryTest.java
@@ -2,7 +2,7 @@ package com.meoguri.linkocean.domain.profile.command.persistence;
 
 import static com.meoguri.linkocean.domain.bookmark.entity.vo.Category.*;
 import static com.meoguri.linkocean.domain.bookmark.entity.vo.OpenType.*;
-import static com.meoguri.linkocean.domain.bookmark.entity.vo.ReactionType.*;
+import static com.meoguri.linkocean.domain.profile.command.entity.vo.ReactionType.*;
 import static com.meoguri.linkocean.domain.user.entity.vo.OAuthType.*;
 import static java.util.List.of;
 import static org.assertj.core.api.Assertions.*;
